move enum ProgressStatus in common
