+++ packages/grant-explorer/src/features/api/types.ts
@@ -57,10 +57,3 @@ export type CartProject = Project & {
   chainId: ChainId;
   amount: string;
 };
-
-export enum ProgressStatus {
-  IS_SUCCESS = "IS_SUCCESS",
-  IN_PROGRESS = "IN_PROGRESS",
-  NOT_STARTED = "NOT_STARTED",
-  IS_ERROR = "IS_ERROR",
-}
