+++ packages/round-manager/src/context/round/RoundContext.tsx
@@ -1,4 +1,5 @@
-import { ProgressStatus, Round } from "../../features/api/types";
+import { Round } from "../../features/api/types";
+import { ProgressStatus } from "common/src/types";
 import React, { createContext, useContext, useEffect, useReducer } from "react";
 import { useWallet } from "../../features/common/Auth";
 import { getRoundById, listRounds } from "../../features/api/round";
