+++ packages/round-manager/src/context/round/UpdateRoundContext.tsx
@@ -1,5 +1,5 @@
 import React, { SetStateAction, createContext, useContext } from "react";
-import { ProgressStatus } from "../../features/api/types";
+import { ProgressStatus } from "common/src/types";
 import { RoundCategory, UpdateRoundParams } from "common/dist/types";
 import { Allo } from "common";
 import { Hex } from "viem";
