+++ packages/round-manager/src/features/program/CreateProgramPage.tsx
@@ -14,7 +14,8 @@ import Footer from "common/src/components/Footer";
 import ProgressModal from "common/src/components/ProgressModal";
 import { datadogLogs } from "@datadog/browser-logs";
 import ErrorModal from "common/src/components/ErrorModal";
-import { ProgressStatus, ProgressStep } from "../api/types";
+import { ProgressStep } from "../api/types";
+import { ProgressStatus } from "common/src/types";
 import { CreateProgramState, useCreateProgram } from "./useCreateProgram";
 import ReactTooltip from "react-tooltip";
 import { CHAINS } from "../api/utils";
