+++ packages/round-manager/src/features/round/FundContract.tsx
@@ -9,7 +9,8 @@ import ReactTooltip from "react-tooltip";
 import { useAccount, useBalance, useNetwork } from "wagmi";
 import { errorModalDelayMs } from "../../constants";
 import { useFundContract } from "../../context/round/FundContractContext";
-import { ProgressStatus, Round } from "../api/types";
+import { Round } from "../api/types";
+import { ProgressStatus } from "common/src/types";
 import { getTxExplorerForContract } from "../api/utils";
 import ConfirmationModal from "../common/ConfirmationModal";
 import ErrorModal from "common/src/components/ErrorModal";
