+++ packages/round-manager/src/features/round/ReclaimFunds.tsx
@@ -7,7 +7,8 @@ import { useNavigate } from "react-router-dom";
 import { useBalance } from "wagmi";
 import { errorModalDelayMs } from "../../constants";
 import { useReclaimFunds } from "../../context/round/ReclaimFundsContext";
-import { ProgressStatus, Round } from "../api/types";
+import { Round } from "../api/types";
+import { ProgressStatus } from "common/src/types";
 import ConfirmationModal from "../common/ConfirmationModal";
 import ErrorModal from "common/src/components/ErrorModal";
 import ProgressModal from "common/src/components/ProgressModal";
