+++ packages/round-manager/src/features/round/ViewRoundResults/ViewRoundResults.tsx
@@ -11,7 +11,8 @@ import { useDropzone } from "react-dropzone";
 import { classNames } from "common";
 import { Button } from "common/src/styles";
 import { useDebugMode, useRoundMatchingFunds } from "../../../hooks";
-import { ProgressStatus, ProgressStep, Round } from "../../api/types";
+import { ProgressStep, Round } from "../../api/types";
+import { ProgressStatus } from "common/src/types";
 import { LoadingRing, Spinner } from "../../common/Spinner";
 import { stringify } from "csv-stringify/sync";
 import { Input } from "csv-stringify/lib";
