+++ packages/round-manager/src/features/round/ViewRoundSettings.tsx
@@ -37,7 +37,8 @@ import { maxDateForUint256 } from "../../constants";
 import { useRoundById } from "../../context/round/RoundContext";
 import { useUpdateRound } from "../../context/round/UpdateRoundContext";
 import { getPayoutTokenOptions, payoutTokens } from "../api/payoutTokens";
-import { ProgressStatus, ProgressStep, Round } from "../api/types";
+import { ProgressStep, Round } from "../api/types";
+import { ProgressStatus } from "common/src/types";
 import { CHAINS, SupportType } from "../api/utils";
 import ConfirmationModal from "../common/ConfirmationModal";
 import ErrorModal from "common/src/components/ErrorModal";
