+++ packages/round-manager/src/features/round/__tests__/ViewRoundResults.test.tsx
@@ -13,7 +13,8 @@ import {
   wrapWithReadProgramContext,
   wrapWithRoundContext,
 } from "../../../test-utils";
-import { ProgressStatus, Round } from "../../api/types";
+import { Round } from "../../api/types";
+import { ProgressStatus } from "common/src/types";
 import ViewRoundPage from "../ViewRoundPage";
 
 jest.mock("common", () => ({
